requirements.txt
annotated-types==0.7.0
anyio==4.11.0
apify_client==1.6.3
apify_shared==1.1.2
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cryptography==44.0.0
deprecation==2.1.0
distro==1.9.0
Flask==3.0.0
Flask-Cors==4.0.0
gotrue==2.12.4
gunicorn==21.2.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.27.2
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
MarkupSafe==3.0.3
openai==2.8.1
packaging==25.0
postgrest==0.18.0
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
realtime==2.24.0
requests==2.31.0
six==1.17.0
sniffio==1.3.1
storage3==0.9.0
StrEnum==0.4.15
supabase==2.10.0
supafunc==0.7.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
